+++ Orm/Xtensive.Orm/Orm/Configuration/Options/Interfaces/INamedOptionsCollectionElement.cs
@@ -1,8 +1,6 @@
-﻿// Copyright (C) 2007-2022 Xtensive LLC.
+// Copyright (C) 2024 Xtensive LLC.
 // This code is distributed under MIT license terms.
 // See the License.txt file in the project root for more information.
-// Created by: Dmitri Maximov
-// Created:    2007.08.03
 
 namespace Xtensive.Orm.Configuration.Options
 {
